Create FairRations.cpp
You are the benevolent ruler of Rankhacker Castle, and today you’re distributing bread. Your subjects are in a line, and some of them already have some loaves. Times are hard and your castle’s food stocks are dwindling, so you must distribute as few loaves as possible according to the following rules:

Every time you give a loaf of bread to some person , you must also give a loaf of bread to the person immediately in front of or behind them in the line (i.e., persons i+1i+1 or i-1i−1).
After all the bread is distributed, each person must have an even number of loaves.
Given the number of loaves already held by each citizen, find and print the minimum number of loaves you must distribute to satisfy the two rules above. If this is not possible, print NO.

For example, the people in line have loaves B=[4,5,6,7]B=[4,5,6,7]. We can first give a loaf to i=3i=3 and i=4i=4 so B=[4,5,7,8]B=[4,5,7,8]. Next we give a loaf to i=2i=2 and i=3i=3 and have B=[4,6,8,8]B=[4,6,8,8] which satisfies our conditions. We had to distribute 44 loaves.

Input Format
The first line contains an integer NN, the number of subjects in the bread line.

The second line contains NN space-separated integers B[i]B[i].

Constraints
2 \leq N \leq 10002≤N≤1000
1 \leq B[i] \leq 101≤B[i]≤10, where 1 \leq i \leq N1≤i≤N
Output Format
Print a single integer that denotes the minimum number of loaves that must be distributed so that every person has an even number of loaves. If it’s not possible to do this, print NO.

Solutions
Wait! Have you challenged yourself with this problem? If yes, click here to show the solution.
First of all, we can observe that the parity of the sum of all the loaves will never change because we always distribute 2 loaves at a time. This means that if the initial sum of the array is odd, we can never satisfy the rule consisting in having each person with an even number of loaves. This is just an observation that will be useful later. We do not need to check if the sum is odd in advance.

The strategy to solve the algorithm could be to simply iterate the array and add one loaf to any person holding an odd number of loaves and to the following one. 
At the end of the scan, if the last person holds an even number of loaves then the algorithm has distributed the minimum amount of loaves and at the same time all the people have an even number of loaves. On the other hand, if the last person has an odd number of loaves then it means the problem could not be solved and the answer is “NO”.